+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Remote logging enabled without requiring code changes
 - If its not a Capacitor or Cordova app then expand Scripts
+- Fix detection of Capacitor projects that do not use @capacitor/core
+- Fix color logging for vite projects
+- Provide recommendation to projects whose ionic.config.json is malformed
 
 ### Version 1.27.1