+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - If its not a Capacitor or Cordova app then expand Scripts
 - Fix detection of Capacitor projects that do not use @capacitor/core
 - Fix color logging for vite projects
+- Fix project statistics for projects folders that have spaces
 - Provide recommendation to projects whose ionic.config.json is malformed
 
 ### Version 1.27.1